backend/requirements.txt
fastapi
uvicorn[standard]
openai>=1.0.0
langgraph>=0.2.0
pydantic
python-dotenv
supabase
requests
typing-extensions
python-multipart
rapidfuzz
email-validator
beautifulsoup4
selenium
webdriver-manager
lxml
html5lib
sqlalchemy
psycopg2-binary
httpx 
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
APScheduler>=3.10.4
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
sentence-transformers>=2.2.0 